Add a simple mnist api server
