+++ Makefile
@@ -20,3 +20,9 @@ utest:
 
 cov:
 	open htmlcov/index.html
+
+serving:
+	PYTHONPATH=src uvicorn src.backend:app --reload
+
+inference:
+	curl -X POST "http://127.0.0.1:8000/predict" -H  "accept: application/json" -H  "Content-Type: multipart/form-data" -F "image=@mnist_sample.jpg;type=image/jpeg"
