+++ README.md
@@ -1,52 +1,30 @@
 # MNIST API Server
 MNIST inference server w/ FastAPI
 
+## Prerequisites
+```bash
+$ make env      # create anaconda environment
+$ make setup    # initial setup for the project
+```
+
+## How to Play
+```bash
+$ make serving      # start up the server
+$ make inference    # send a sample image file to the server
+```
+
 ## Training Conv Net with MNIST
 1. Go to `src/ml`
 2. Run `python train.py`
 
-## References
-- https://github.com/pytorch/examples/blob/master/mnist/main.py
-
-
-## File Structure
+## For Developers
 ```bash
-.
-├── LICENSE
-├── Makefile          # commands
-├── README.md
-├── requirements.txt  # package information
-├── setup.cfg         # configurations for formatting & linting & unit-test
-├── src               # source code location
-└── test
-    └── utest         # unit tests location
-```
-
-## Commands
-```bash
-$ make env      # create anaconda environment
-$ make setup    # initial setup for the project
 $ make format   # format python scripts
 $ make lint     # lint python scripts
 $ make utest    # run unit tests
 $ make cov      # open coverage report (after `make utest`)
 ```
 
-## Configurations
-`setup.cfg` states all configurations for formatting & linting & unit-test.
-
-## Verifications
-- per commit: pre-commit hook runs formatting and linting.
-- per pull-request: GitHub Actions check formatting, linting, and unit-test results.
-
-## Recommended Repository Settings
-#### Restriction on multi-commit pushes
-`Settings` -> `General` -> `Merge botton` -> `Allow squash merging` ONLY
-<img width="796" src="https://user-images.githubusercontent.com/14961526/152031596-a329a74c-add7-4d1c-ada5-d0279da16195.png">
-
-#### Branch Protection Rules
-`Settings` -> `Branches` -> `Branch protection rules` -> `Add rule`
-- Branch name pattern: `main`
-- Require a pull request before merging & Require approvals
-- Require status checks to pass before merging & Require branches to be up to date before merging
-- Include administrators
+## References
+- https://github.com/pytorch/examples/blob/master/mnist/main.py
+- https://github.com/KiLJ4EdeN/fastapi_tf-keras_example
