+++ requirements.txt
@@ -27,3 +27,8 @@ torchvision     == 0.9.1
 
 # utils
 tqdm            == 4.62.3
+
+# backend
+fastapi             == 0.73.0
+uvicorn[standard]   == 0.17.4
+python-multipart    == 0.0.5
